Remove fastify-bcrypt dependency and related code; update user repository interface
