+++ package.json
@@ -31,7 +31,6 @@
     "drizzle-zod": "^0.5.1",
     "fastify": "^5.0.0",
     "fastify-api-reference": "link:@types/@scalar/fastify-api-reference",
-    "fastify-bcrypt": "^1.0.1",
     "fastify-plugin": "^5.0.1",
     "fastify-type-provider-zod": "^4.0.2",
     "fastify-zod": "^1.4.0",
