+++ src/bootstrap.ts
@@ -6,10 +6,8 @@ import mutipart from "@fastify/multipart";
 import limiter from "@fastify/rate-limit";
 import swagger from "@fastify/swagger";
 import scalar from "@scalar/fastify-api-reference";
-import { fastifyBcrypt } from "fastify-bcrypt";
 import { serializerCompiler, validatorCompiler } from "fastify-type-provider-zod";
 
-import { env } from "@/env";
 import { swaggerConfig, swaggerScalarConfig } from "@/open-api";
 import authenticate from "@/plugins/authenticate";
 import { corsConfig, rateLimiterConfig } from "@/utils/http";
@@ -17,7 +15,6 @@ import { corsConfig, rateLimiterConfig } from "@/utils/http";
 import { ErrorHandlers } from "./handlers/error/error-handlers";
 import database from "./plugins/db";
 import cache from "./plugins/redis-cache";
-import { bcryptSaltConfig } from "./utils/jwt";
 
 export async function bootstrapServerPlugins(server: FastifyInstance) {
   try {
@@ -28,7 +25,6 @@ export async function bootstrapServerPlugins(server: FastifyInstance) {
     await server.register(cache);
     await server.register(database);
     await server.register(limiter, { redis: server.cache, ...rateLimiterConfig });
-    await server.register(fastifyBcrypt, bcryptSaltConfig);
     await server.register(cors, corsConfig);
     await server.register(clerkPlugin);
