+++ src/handlers/user/user-schemas.ts
@@ -4,21 +4,14 @@ import { z } from "zod";
 import * as schema from "@/db/schema";
 import { commonHttpResponseSchema } from "@/utils/http";
 
-import type { signinUserSchema } from "../auth/auth-schemas";
-
-import { signupUserSchema } from "../auth/auth-schemas";
-
 export const foundUserSchema = createSelectSchema(schema.users);
-
-export const updateUserSchema = signupUserSchema.partial();
+export const updateUserSchema = foundUserSchema.partial();
 
 export type TUserToUpdate = z.infer<typeof updateUserSchema>;
-export type TUserToCreate = z.infer<typeof signupUserSchema>;
-export type TCandidateUser = z.infer<typeof signinUserSchema>;
 export type TFoundUserResult = z.infer<typeof foundUserSchema>;
 
 export const userDetailsResponseSchema = commonHttpResponseSchema.extend({
-  data: foundUserSchema.omit({ password: true }),
+  data: foundUserSchema,
 });
 
 export const userDeletedResponseSchema = z.object({
